Add the ability to zoom/unzoom the current pane
This means making the current pane take up the majority of the current
working space and shrinking all other panes down to very minimal size.
This allows your to focus working on one pane without destroying the
others. I find it very useful on a screen with a smaller resolution