Allow marking certain layouts as favourite
This commit allows the user to mark certain layouts as favourite
for use with `new_window_with_favourite_layout` command, which,
much like `new_window_with_current_layout`, opens a new window;
the difference is that said marked layout is used as a base
instead of the current one.

If there are multiple layouts marked as favourite, only the first
(per natural ordering in `saved_layouts`) is chosen.